requirements.txt
PyDispatcher==2.0.3
WebOb==1.2.3
WebTest==2.0.9
Whoosh==2.5.4
beautifulsoup4==4.3.2
bottle==0.12-dev
coverage==3.7
distribute==0.7.3
nose==1.3.0
pyxdg==0.25
redis==2.8.0
six==1.4.1
waitress==0.8.7
